--- src/main/java/com/example/study/product/command/application/CouponProductDto.java
@@ -1,7 +1,0 @@
-package com.example.study.product.command.application;
-
-public record CouponProductDto(
-        int discountPrice,
-
-        int effectiveDay
-){}
