--- src/main/java/com/example/study/product/command/application/DeliveryProductDto.java
@@ -1,7 +1,0 @@
-package com.example.study.product.command.application;
-
-public record DeliveryProductDto(
-        int fee,
-
-        int weight
-){}
