--- src/main/java/com/example/study/product/command/application/ProductTagDto.java
@@ -1,8 +1,0 @@
-package com.example.study.product.command.application;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record ProductTagDto(
-        @NotBlank(message = "상품 태그 이름은 필수입니다.")
-        String tagName
-){}
