+++ src/main/java/com/example/study/product/command/application/product/CouponProductDto.java
@@ -1,0 +1,7 @@
+package com.example.study.product.command.application.product;
+
+public record CouponProductDto(
+        int discountPrice,
+
+        int effectiveDay
+){}
