+++ src/main/java/com/example/study/product/command/application/product/DeliveryProductDto.java
@@ -1,0 +1,7 @@
+package com.example.study.product.command.application.product;
+
+public record DeliveryProductDto(
+        int fee,
+
+        int weight
+){}
