+++ src/main/java/com/example/study/product/command/application/review/ReviewConstant.java
@@ -1,0 +1,5 @@
+package com.example.study.product.command.application.review;
+
+public class ReviewConstant {
+    public static final int pagingSize = 10;
+}
