+++ src/main/java/com/example/study/product/command/application/review/ReviewNotFoundException.java
@@ -1,0 +1,11 @@
+package com.example.study.product.command.application.review;
+
+import com.example.study.common.CustomException;
+import org.springframework.http.HttpStatus;
+
+public class ReviewNotFoundException extends CustomException {
+
+    public ReviewNotFoundException() {
+        super("해당 리뷰정보를 찾을 수 없습니다.", HttpStatus.NOT_FOUND);
+    }
+}
