+++ src/main/java/com/example/study/product/command/domain/review/ReviewRepository.java
@@ -1,0 +1,11 @@
+package com.example.study.product.command.domain.review;
+
+import java.util.Optional;
+
+public interface ReviewRepository {
+    Review save(Review review);
+
+    Optional<Review> findByReviewId(Long reviewId);
+
+    void delete(Review review);
+}
