--- src/main/java/com/example/study/review/command/application/InvalidReviewAuthenticationException.java
@@ -1,11 +1,0 @@
-package com.example.study.review.command.application;
-
-import com.example.study.common.CustomException;
-import org.springframework.http.HttpStatus;
-
-public class InvalidReviewAuthenticationException extends CustomException {
-
-    public InvalidReviewAuthenticationException() {
-        super("자신이 작성한 리뷰만 수정 할 수 있습니다.", HttpStatus.NOT_FOUND);
-    }
-}
