--- src/main/java/com/example/study/review/command/application/ReviewConstant.java
@@ -1,5 +1,0 @@
-package com.example.study.review.command.application;
-
-public class ReviewConstant {
-    public static final int pagingSize = 10;
-}
